Refactor: Apply MVI (2)
