+++ presentation/src/main/kotlin/io/github/jisungbin/gitmessengerbot/mvi/MviBaseState.kt
@@ -2,13 +2,16 @@
  * GitMessengerBot © 2021 지성빈 & 구환. all rights reserved.
  * GitMessengerBot license is under the GPL-3.0.
  *
- * [MviBaseState.kt] created by Ji Sungbin on 21. 9. 18. 오후 8:56
+ * [MviBaseState.kt] created by Ji Sungbin on 21. 9. 19. 오후 4:47
  *
  * Please see: https://github.com/GitMessengerBot/GitMessengerBot-Android/blob/master/LICENSE.
  */
 
 package io.github.jisungbin.gitmessengerbot.mvi
 
-sealed class MviBaseSideEffect {
-    data class Toast(val message: String) : MviBaseSideEffect()
+interface MviBaseState {
+    val loading: Boolean
+    val exception: Exception?
+
+    fun isException() = exception == null
 }
